// src/Libs/interfaces.ts
export interface IFormUploader {
  uploadPreset: string;
}

export interface ICloudinaryResponse {
  asset_id: string;
  public_id: string;
  version: number;
  version_id: string;
  signature: string;
  width: number;
  heigth: number;
  format: string;
  resource_type: string;
  created_at: string;
  tags: object;
  bytes: number;
  type: string;
  etag: string;
  placeholder: boolean;
  url: string;
  secure_url: string;
  folder: string;
  access_mode: string;
  original_filename: string;
}